Refactor experiment metadata config into separate component
